fixed issue where types were not searchable
